Replace switch statement with an object literal.
Technically, there is little difference because a lookup is going to have to happen
anyway, but I believe this option is cleaner and shorter.

This commit also removes the exception being thrown in the default case. Now, a different
exception will be thrown in the default case (the engine will attempt to call 'undefined'),
but this case is not hardcoded in the code because, again, that usage isn't documented
in the spec, so it shouldn't be used in the first place.