Remove old tests.
They checked the RegExp behaviour. I believe the problem that regularity is
solving is the correct formation of RegExps, and therefore that is what must
be tested. I think there is little value in testing that the returned RegExp
works properly, since that is assured by the spec and the engine's conformance
to it.